refactor(profile): rearanged directory structure
