+++ lib/views/pages/profile/profile_page.dart
@@ -2,11 +2,11 @@ import "package:flutter/material.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/utils/ui/circular_value.dart";
 import "package:proxima/viewmodels/profile_view_model.dart";
+import "package:proxima/views/pages/profile/components/profile_app_bar.dart";
 import "package:proxima/views/pages/profile/components/profile_badge.dart";
-import "package:proxima/views/pages/profile/components/profile_info_card.dart";
-import "package:proxima/views/pages/profile/components/profile_info_column.dart";
-import "package:proxima/views/pages/profile/components/profile_info_row.dart";
-import "package:proxima/views/pages/profile/profile_app_bar.dart";
+import "package:proxima/views/pages/profile/info_cards/profile_info_card.dart";
+import "package:proxima/views/pages/profile/info_cards/profile_info_column.dart";
+import "package:proxima/views/pages/profile/info_cards/profile_info_row.dart";
 
 /// This widget is used to display the profile page
 /// It contains the user info, centauri points, badges, posts and comments
